Move code to src/ subdir
To separate source code from other files (tests, doc)
