+++ src/generate-function.js
