+++ src/jaystring.js
