+++ src/pointer.js
