+++ test/schema-path.js
@@ -1,6 +1,6 @@
 const tape = require('tape')
 const validator = require('../')
-const { get } = require('../pointer')
+const { get } = require('../src/pointer')
 
 const target = Symbol('target')
